requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=1.0.0
tavily-python>=0.3.0
anthropic>=0.20.0
notion-client>=2.2.0
pyspellchecker>=0.7.0
lxml>=5.0.0
